+++ src/hooks/level_browser_layer.cpp
@@ -14,8 +14,11 @@ void HookedLevelBrowserLayer::setupLevelBrowser(cocos2d::CCArray* p0) {
     if (!p0 || !nm.established()) return;
 
     std::vector<LevelId> levelIds;
-    for (auto* level : CCArrayExt<GJGameLevel*>(p0)) {
-        if (level->m_levelType >= GJLevelType::Saved) {
+    for (auto* level_ : CCArrayExt<CCObject*>(p0)) {
+        if (!typeinfo_cast<GJGameLevel*>(level_)) continue;
+
+        auto* level = static_cast<GJGameLevel*>(level_);
+        if (isValidLevelType(level->m_levelType)) {
             levelIds.push_back(level->m_levelID);
         }
     }
@@ -33,13 +36,16 @@ void HookedLevelBrowserLayer::setupLevelBrowser(cocos2d::CCArray* p0) {
 }
 
 void HookedLevelBrowserLayer::refreshPagePlayerCounts() {
-    for (auto* cell : CCArrayExt<LevelCell*>(m_list->m_listView->m_tableView->m_contentLayer->getChildren())) {
-        if (cell->m_level->m_levelType < GJLevelType::Saved) continue;
+    for (auto* cell_ : CCArrayExt<CCNode*>(m_list->m_listView->m_tableView->m_contentLayer->getChildren())) {
+        if (!typeinfo_cast<LevelCell*>(cell_)) continue;
+        auto* cell = static_cast<GlobedLevelCell*>(cell_);
+
+        if (!isValidLevelType(cell->m_level->m_levelType)) continue;
 
         if (m_fields->levels.contains(cell->m_level->m_levelID)) {
-            static_cast<GlobedLevelCell*>(cell)->updatePlayerCount(m_fields->levels.at(cell->m_level->m_levelID));
+            cell->updatePlayerCount(m_fields->levels.at(cell->m_level->m_levelID));
         } else {
-            static_cast<GlobedLevelCell*>(cell)->updatePlayerCount(-1);
+            cell->updatePlayerCount(-1);
         }
     }
 }
@@ -48,9 +54,11 @@ void HookedLevelBrowserLayer::updatePlayerCounts(float) {
     auto& nm = NetworkManager::get();
     if (nm.established()) {
         std::vector<LevelId> levelIds;
-        for (auto* cell : CCArrayExt<LevelCell*>(m_list->m_listView->m_tableView->m_contentLayer->getChildren())) {
-            if (cell->m_level->m_levelType >= GJLevelType::Saved) {
-                levelIds.push_back(cell->m_level->m_levelID);
+        for (auto* cell_ : CCArrayExt<CCNode*>(m_list->m_listView->m_tableView->m_contentLayer->getChildren())) {
+            if (auto* cell = typeinfo_cast<LevelCell*>(cell_)) {
+                if (isValidLevelType(cell->m_level->m_levelType)) {
+                    levelIds.push_back(cell->m_level->m_levelID);
+                }
             }
         }